+++ GameServer/Server/CallGS/Handlers/Tower/ClimbTowerLogic_GetReward.cs
@@ -242,7 +242,7 @@ public class ClimbTowerLogic_GetReward : ICallGSHandler
 
     private static BaseGameItemInfo? AddOtherItem(InventoryData inventory, uint genre, uint detail, uint particular, uint level, uint count)
     {
-        var templateId = (uint)GameResourceTemplateId.FromGdpl(genre, detail, particular, level);
+        var templateId = GameResourceTemplateId.FromGdpl(genre, detail, particular, level);
         if (!GameData.OtherItemData.TryGetValue(templateId, out var otherItem))
             return null;
