+++ GameServer/Server/CallGS/Handlers/VirCapture/VirCapture_GetLevelAward.cs
@@ -258,7 +258,7 @@ public class VirCapture_GetLevelAward : ICallGSHandler
 
     private static BaseGameItemInfo? AddOtherItem(InventoryData inventory, uint genre, uint detail, uint particular, uint level, uint count)
     {
-        var templateId = (uint)GameResourceTemplateId.FromGdpl(genre, detail, particular, level);
+        var templateId = GameResourceTemplateId.FromGdpl(genre, detail, particular, level);
         if (!GameData.OtherItemData.TryGetValue(templateId, out var otherItem))
             return null;
